feat: Version the CRD types
