+++ rust/operator-binary/src/backend/dynamic.rs
@@ -15,10 +15,7 @@ use super::{
     pod_info::{PodInfo, SchedulingPodInfo},
     tls,
 };
-use crate::{
-    crd::{self, SecretClass},
-    utils::Unloggable,
-};
+use crate::{crd::v1alpha1, utils::Unloggable};
 
 pub struct DynError(Box<dyn SecretBackendError>);
 
@@ -129,10 +126,10 @@ impl SecretBackendError for FromClassError {
 
 pub async fn from_class(
     client: &stackable_operator::client::Client,
-    class: SecretClass,
+    class: v1alpha1::SecretClass,
 ) -> Result<Box<Dynamic>, FromClassError> {
     Ok(match class.spec.backend {
-        crd::SecretClassBackend::K8sSearch(crd::K8sSearchBackend {
+        v1alpha1::SecretClassBackend::K8sSearch(v1alpha1::K8sSearchBackend {
             search_namespace,
             trust_store_config_map_name,
         }) => from(super::K8sSearch {
@@ -140,7 +137,7 @@ pub async fn from_class(
             search_namespace,
             trust_store_config_map_name,
         }),
-        crd::SecretClassBackend::AutoTls(crd::AutoTlsBackend {
+        v1alpha1::SecretClassBackend::AutoTls(v1alpha1::AutoTlsBackend {
             ca,
             additional_trust_roots,
             max_certificate_lifetime,
@@ -153,11 +150,11 @@ pub async fn from_class(
             )
             .await?,
         ),
-        crd::SecretClassBackend::CertManager(config) => from(super::CertManager {
+        v1alpha1::SecretClassBackend::CertManager(config) => from(super::CertManager {
             client: Unloggable(client.clone()),
             config,
         }),
-        crd::SecretClassBackend::KerberosKeytab(crd::KerberosKeytabBackend {
+        v1alpha1::SecretClassBackend::KerberosKeytab(v1alpha1::KerberosKeytabBackend {
             realm_name,
             kdc,
             admin,
@@ -185,14 +182,14 @@ pub enum FromSelectorError {
     #[snafu(display("failed to get {class}"))]
     GetSecretClass {
         source: stackable_operator::client::Error,
-        class: ObjectRef<SecretClass>,
+        class: ObjectRef<v1alpha1::SecretClass>,
     },
 
     #[snafu(display("failed to initialize backend for {class}"))]
     FromClass {
         #[snafu(source(from(FromClassError, Box::new)))]
         source: Box<FromClassError>,
-        class: ObjectRef<SecretClass>,
+        class: ObjectRef<v1alpha1::SecretClass>,
     },
 }
 
@@ -220,7 +217,7 @@ pub async fn from_selector(
 ) -> Result<Box<Dynamic>, FromSelectorError> {
     let class_ref = || ObjectRef::new(&selector.class);
     let class = client
-        .get::<SecretClass>(&selector.class, &())
+        .get::<v1alpha1::SecretClass>(&selector.class, &())
         .await
         .with_context(|_| from_selector_error::GetSecretClassSnafu { class: class_ref() })?;
     from_class(client, class)
