+++ rust/operator-binary/src/main.rs
@@ -15,14 +15,17 @@ use grpc::csi::v1::{
     controller_server::ControllerServer, identity_server::IdentityServer, node_server::NodeServer,
 };
 use stackable_operator::{
-    CustomResourceExt,
+    YamlSchema,
     cli::{CommonOptions, ProductOperatorRun},
+    shared::yaml::SerializeOptions,
     telemetry::Tracing,
 };
 use tokio::signal::unix::{SignalKind, signal};
 use tokio_stream::wrappers::UnixListenerStream;
 use tonic::transport::Server;
 use utils::{TonicUnixStream, uds_bind_private};
+
+use crate::crd::{SecretClass, TrustStore};
 
 mod backend;
 mod crd;
@@ -69,8 +72,10 @@ async fn main() -> anyhow::Result<()> {
     let opts = Opts::parse();
     match opts.cmd {
         stackable_operator::cli::Command::Crd => {
-            crd::SecretClass::print_yaml_schema(built_info::PKG_VERSION)?;
-            crd::TrustStore::print_yaml_schema(built_info::PKG_VERSION)?;
+            SecretClass::merged_crd(crd::SecretClassVersion::V1Alpha1)?
+                .print_yaml_schema(built_info::PKG_VERSION, SerializeOptions::default())?;
+            TrustStore::merged_crd(crd::TrustStoreVersion::V1Alpha1)?
+                .print_yaml_schema(built_info::PKG_VERSION, SerializeOptions::default())?;
         }
         stackable_operator::cli::Command::Run(SecretOperatorRun {
             csi_endpoint,
